raw conversion of SelfInitializedFakes specs
